+++ Sources/AccioKit/Globals/Constants.swift
@@ -19,18 +19,4 @@ enum Constants {
 
         return FileManager.userCacheDirUrl.appendingPathComponent("Accio/Cache").path
     }
-
-    static var swiftVersion: String {
-        #if swift(>=6.0)
-            return "Swift-6.0"
-        #elseif swift(>=5.2)
-            return "Swift-5.2"
-        #elseif swift(>=5.1)
-            return "Swift-5.1"
-        #elseif swift(>=5.0)
-            return "Swift-5.0"
-        #else
-            return "Swift-4.2"
-        #endif
-    }
 }