+++ Sources/AccioKit/Services/XcodeProjectIntegrationService.swift
@@ -85,6 +85,16 @@ final class XcodeProjectIntegrationService {
         try bash("rm -rf \(dependenciesPath)")
     }
 
+    func copy(cachedFrameworkProducts: [FrameworkProduct]) throws {
+        for platform in Platform.allCases {
+            let dependenciesPath = "\(workingDirectory)/\(Constants.dependenciesPath)/\(platform)"
+            let cachedFrameworkProducts = cachedFrameworkProducts.filter { $0.platformName == platform.rawValue }
+
+            print("Copying \(cachedFrameworkProducts.count) cached build products to Dependencies folder for platform \(platform.rawValue).", level: .info)
+            try copy(frameworkProducts: cachedFrameworkProducts, to: dependenciesPath)
+        }
+    }
+
     func updateDependencies(of appTarget: AppTarget, for platform: Platform, with frameworkProducts: [FrameworkProduct]) throws {
         let dependenciesPlatformPath = "\(workingDirectory)/\(Constants.dependenciesPath)/\(platform.rawValue)"
         let copiedFrameworkProducts: [FrameworkProduct] = try copy(frameworkProducts: frameworkProducts, of: appTarget, to: dependenciesPlatformPath)
@@ -93,7 +103,11 @@ final class XcodeProjectIntegrationService {
 
     private func copy(frameworkProducts: [FrameworkProduct], of appTarget: AppTarget, to targetPath: String) throws -> [FrameworkProduct] {
         print("Copying build products of target '\(appTarget.targetName)' into folder '\(Constants.dependenciesPath)' ...", level: .info)
+        return try copy(frameworkProducts: frameworkProducts, to: targetPath)
+    }
 
+    @discardableResult
+    private func copy(frameworkProducts: [FrameworkProduct], to targetPath: String) throws -> [FrameworkProduct] {
         try bash("mkdir -p '\(targetPath)'")
         var copiedFrameworkProducts: [FrameworkProduct] = []
 
@@ -112,8 +126,9 @@ final class XcodeProjectIntegrationService {
             try bash("cp -R '\(frameworkProduct.frameworkDirPath)' '\(frameworkDirPath)'")
             try bash("cp -R '\(frameworkProduct.symbolsFilePath)' '\(symbolsFilePath)'")
 
-            let frameworkProduct = FrameworkProduct(frameworkDirPath: frameworkDirPath, symbolsFilePath: symbolsFilePath)
+            let frameworkProduct = FrameworkProduct(frameworkDirPath: frameworkDirPath, symbolsFilePath: symbolsFilePath, commitHash: frameworkProduct.commitHash)
             try frameworkProduct.cleanupRecursiveFrameworkIfNeeded()
+            try verifyBundleVersion(of: frameworkProduct)
 
             copiedFrameworkProducts.append(frameworkProduct)
         }
@@ -226,5 +241,20 @@ final class XcodeProjectIntegrationService {
         }
 
         try projectFile.write(path: Path(xcodeProjectPath), override: true)
+    }
+
+    private func verifyBundleVersion(of product: FrameworkProduct) throws {
+        let plistURL = product.frameworkDirUrl.appendingPathComponent("Info.plist")
+        let data = try Data(contentsOf: plistURL)
+        var format: PropertyListSerialization.PropertyListFormat = .binary
+        var plist = try PropertyListSerialization.propertyList(from: data, options: [.mutableContainersAndLeaves], format: &format) as! [String: Any]
+
+        if plist["CFBundleVersion"] == nil {
+            print("CFBundleVersion of framework \(product.libraryName) was not specified and will be set to 1", level: .warning)
+            plist["CFBundleVersion"] = "1"
+
+            let data = try PropertyListSerialization.data(fromPropertyList: plist, format: format, options: 0)
+            try data.write(to: plistURL, options: .atomic)
+        }
     }
 }